requirements.txt
Flask==2.3.3
phe==1.5.0
requests
python-dotenv
bleak>=0.20.0